fix(packaging): the install one-liner runs under sh [no-changelog]
The published command pipes the script into `sh`, which ignores the
shebang, and `sh` is dash on Debian and Ubuntu. Two lines refused there.
`set -o pipefail` is not in dash before 0.5.12 and not in Ubuntu's build of
it at all, so the script exited at line 9 on every Ubuntu tested. `trap ...
RETURN` is bash's alone and refused by every dash, including the ones that
take pipefail, so Debian reached the first download and stopped. Measured
under dash on Debian trixie and Ubuntu 20.04, 22.04 and 24.04, where the
script now installs the command, the app, the launcher entry and all four
icons and leaves no temporary directory behind.

Dropping `pipefail` costs a guarantee over the one pipeline in the script,
the release lookup. An empty result is checked for by name a few lines
below, which stops the install with a sentence rather than an abrupt exit,
and a test holds that path.

The test ran the script under `bash`, which is how both bashisms reached
the published command with every check green, so it runs it the way the
published command does — preferring dash where it is installed, since
`/bin/sh` is bash on some distributions and would let the next one through.
